+++ src/NetStandard/E.Standard.WebGIS.Tools/Georeferencing/Image/PrintImageService.cs
@@ -52,6 +52,7 @@ internal class PrintImageService : IStaticOverlayService, IPrintableService
     public string ID { get; set; }
 
     public float Opacity { get; set; }
+    public float OpacityFactor { get; set; }
 
     public bool CanBuffer => false;