+++ src/NetStandard/E.Standard.WebMapping.GeoServices/OGC/WFS/WfsService.cs
@@ -93,6 +93,7 @@ public class WfsService : IMapService, IServiceSupportedCrs, IServiceCopyrightIn
 
         }
     }
+    public float OpacityFactor { get; set; } = 1f;
 
     public bool ShowInToc { get; set; }